use inotify_simple for FS watching
The existing watchdog is pretty awkward to use proceduraly,
and adds unnecessary abstraction. What we want is installing
inotify on new dirs and removing inotify once complete, so
we don't need hundreds of thousands watches on the entire tree.

Use a simple inotify wrapper, and code the logic up manually.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>