tests: patch: build_clang_rust: Build with rust enabled
Add a build test, which is based on clang/llvm, but additionally
enables Rust.

The test is a NOP is the patch does not contain and .rs files, in
order to keep the additional load on the build servers to a
minimum. This does however mean a change to C code could break the
rust code, and would not be noted.

Test that the build system has a usable version of the rust toolchain,
since it changes from release to release.

Create a custom kernel configuration, based on allmodconfig, but with
some options disabled in order to allow Rust to be
enabled. Additionally enabled the Rust PHYLIB binding and the one Rust
PHY driver in the tree.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>