cc_maintainers: add staleness filter and pull request ignore
Introduce the ability to "log to stdout" by returning 3
values from python tests. Log info from cc_maintainers.

Check if we heard from people who are blamed in the last
3 years, and ignore them if we didn't. This is implemented
as a public inbox query so cache the info for 2 weeks to
avoid frequent refreshes. This can be further optimized
to recalculate the age based on existing results for
active folks.

Also don't require known bona fide sub-maintainers to CC
people when co-posting patches with PRs.

Before pushing this change avg runtime for the test
on netdev was 1.4s, we'll see how much this slows us down...

Signed-off-by: Jakub Kicinski <kuba@kernel.org>