remote-monitor: add a simple service which checks if NIPA is running
Add a script which can be run on a laptop or such,
which fetches the status of NIPA from remote URL (the json
which NIPA renders) and shows a pop up when something fails.

Status for netdev can be fetched from:

  https://netdev.bots.linux.dev/static/nipa/systemd.json

the script takes the URL as the first (and only) arg.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>