codegen: generate a diff of user code
We want to move away from committing the user space code in the tree.
We can generate all that code on the fly and put it in .gitignore.
But for reviewing it's useful to see if the series has any impact
on what's generated. So have the codegen check generate a diff.

We need to export BRANCH_BASE and RESULTS_DIR so the script
knows where to write stuff.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>