core: tester: apply the series before trying to run tests
Other series tests operate on patch diffs directly,
so we never had to apply the patches to the tree.
This is no longer true for YNL, we need to apply
the series before testing it.

While at it, remove the fetch from the reset.
It shouldn't be needed, and worst case it will
fetch new remove after we checked if patches
apply and they will no longer apply.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>