README: add acknowledgments
Netronome sponsored the initial version of NIPA (by letting
me work on it while I was employed there). It's quite rare
to be this nice to the Open Source community, especially
for a startup. The repo is moving from Netronome org to our
new linux-netdev GitHub org - add an acknowledgment for
the sponsors of the work.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>