ui: status: deal with missing start times
Right after the branch is created but before branch fetcher
produced the fake "branch creation succeeded" result - we're
in a situation where the brancher executor is missing a result.
We insert an empty result for the missing run.
Normally the start for that inserted result would be the start
of branch... but there is no branch. So we end up setting result
start to 'undefined'. Calling toLocaleString() on undefined will
explode.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>