pw_poller: flip the nesting of test execution
Instead of iterating:

  foreach patch:
    foreach test:
      test.run()

do:

  foreach test:
    foreach patch:
      test.run()

build tests need to put the tree into a fully built state first.
A state where there's nothing to do for an incremental build.
If we don't interleave the tests they can assume the tree is
in whatever state it was when testing the previous patch finished
plus git am next.patch, but be sure that absolutely nothing
else has happened.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>