contest: vm: 'slowdown' option to increase timeout
When this option is set, two different env vars are set:

- KSFT_MACHINE_SLOW: some tests check if the env var is defined (no
  matter the value it is set to), and are then more tolerant with
  issues. No need to set this env var in the 'exports' option everywhere
  needed then.

- kselftest_override_timeout: the kselftest timeout is multiplied by the
  float value linked to this new 'slowdown' option. The kselftest's
  'settings' file is read to get the custom timeout option. If not set,
  the default timeout of 45 seconds is used. Increasing the timeout is
  useful when the environment is "abnormally" slow, e.g. when using a
  bunch of debug kconfig. This is probably better to override the
  timeout instead of setting a huge value supporting such envs, not to
  have a too high value for "normal" environments, loosing the interest
  of "quickly" stopping a stalled test.

Note that the _get_ksft_timeout() method could maybe be moved to
vmksft. I didn't do it as it is not clear to me if vmksft-p will be
merged with vmksft or if the latter will be removed. Anyway, this ksft
specific code is in a separated method. So if we want to move stuff, it
should not be too hard.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>