contest: ui: centralize and redesign the UI "pw reporting" logic
Move the result filtering to nipa.js. Switch to using a new
"ignore-results" table in the JSON file, which allows filtering
on any number of fields.

Note that the actual PW reporter (the Python service) is not
changed yet! It will be changed once the flakes are shaken
out and when time allows...

Signed-off-by: Jakub Kicinski <kuba@kernel.org>