contest: vmksft-p: number vms from 1
VM IDs are what we start from + 1, so the first VM in each
thread will have ID number 2:

vm-start-thr0-2
vm-start-thr1-2
vm-stop-thr0-2
vm-stop-thr1-2

(the "-2" at the end). This is pretty odd. Start from 0, like thread ID.
Note that the initial value of vm_id is the "previous ID", so init to -1.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>