pw_contest: fix handling on unreal in final report, again
We see results getting reported as pass even when some tests
are still pending. This is because for null results we use
UNKNOWN and UNKNOWN is -2, so max() will always pick something
else.

There's another bug here, were we hard-overwrite the result
with Pending. That's not necessary, if there's a failure
(high code) we can report right away.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>