tests: build: make gcc builds avoid rebuilds as well
Starting with commit 762fec308b13 ("tests: build_clang: take advantage
of $FIRST_IN_SERIES") (and subsequent fixes) - clang build test tries
to avoid doing a full rebuild for the "baseline". Do the same in GCC
builds, this can save up to 17% (40% -> 35%) of build time on days with
long series. And it's the long series we care about the most.

The print of output lengths was probably to do later analysis
but the wins were clear enough in the status UI, remove it.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>