html: support dark scheme
While working on an MPTCP version of "flakes", the Dark Reader extension
I used with my browser was poorly rendering the page when used from a
local server.

I realised it was easy to add a dark scheme support by overriding some
colours in a '@media (prefers-color-scheme: dark)' section, in the CSS.
So it will used the dark scheme if the browser prefers a dark scheme.

Note that a text written with the default 'blue' colour is hardly
visible when a dark background is used. So I used the recommended
'#809fff' colour instead, a bit lighter, still blue. It is used with and
without the dark scheme.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>