mailbot: process delay_actions in order of time
We used to delay all actions by the same amount of time.
So the processing check treats the action list as a queue.
Now we delay by 1h or 24h, so we need to make sure the list
is indeed sorted.

Note that the list contains tuples (time, msg obj), but time
should never be equal, so msg objects won't get compared.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>