contest: fetcher: log results and info into PostgreSQL
CouchDB is not ideal in terms of perf for most common
queries. Also record the results in PostgreSQL so we
can migrate to that slowly.

Results are split into "normal" and "full", where normal
is what most UIs want. We can return that precomputed,
and already minimized in terms of transfer size.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>