contest: kunit: add sub-cases
Format the results as the new "l2" thing, with an optional
"results" entry in every case, which will hold sub-cases.
The first level of results is unchanged* (and must contain
the overall "result"), the second layer of results only
have test name and result. No futher nesting.

* There's actually a change in the results with this patch,
because the previous parsing was broken. I _intended_ to descend
recursively to subgroups but it only worked for the first layer
(it tried to go two levels deep when recursing).

Signed-off-by: Jakub Kicinski <kuba@kernel.org>