system-status: log DB size
The DB might get info about subtests when available. It is then
important to monitor the DB size, to quickly take actions if it grows
too fast.

Here, the system-status script will check the size once per day. The
size of the last 40 days are written in the JSON, that could be
displayed in the status page later.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>