contest: kunit: 'namify' the test names
The new subtests have spaces in their name, e.g.

  example.example_params_test.example value 0

It sounds better to avoid these spaces and other unwanted chars, as the
test name is used to identify the test.

This will break the tracking of failed tests, but it has just been
broken a few days ago, probably the good time to do that again.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>