system-status: db: reverse order
The list we get from the DB is reversed, to be able to easily limit it
to 40 entries. To present the data in the right order in the JSON, we
need to reverse them again.

The UI could also reverse them, but it sounds better to write the data
in the right order instead.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>