contest: vmksft-p: nested: support sub-cases format
Similar to commit bf87f3d ("contest: kunit: add sub-cases"), sub-cases
are now stored in a list, attached to the parent's result, using the
'results' key.

The nested tests are then now parsed before adding the parent's results
in the queues, to be able to attach them if any.

By doing that, we should be able to re-enable sub-tests parsing for
selftests supporting them: MPTCP, TC, the ones using kselftest_harness.h,
etc.

Note that it is also easy to add a new option to only store failed tests
if needed: just adding a check before 'tests.append()' in
_parse_nested_tests().

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>