mailbot: support MlEmail sorting
Every now and then some bot blasts a lot of emails at once
and we end up with the same exact dates in the defer queue.
The defer queue has tuples of data and MlEmail so when dates
are the same it tries to compare email instances. We don't
care which one gets to the start of the queue first, but
let's not crash.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>