contest: vm: add cmd to start VM in logs
In the logs, we can see the command used to build the VM, the selftests,
etc. but not the one to run the VM.

The command is now added to 'log_out', which will be dumped later to the
'stdout' file.

Note that we could also move the line adding '> TREE CMD:' to log_out
from tree_cmd() to tree_popen(), but that will also add the
decoded_stacktrace.sh command to log_out, which will not contain the
output of this command as it is redirected to another file. So maybe
better not to modify 'tree_popen()' I suppose.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>