tests: build: run baseline build for Kconfig updates
Some Kconfig updates may result in .config changes which may
in turn alter the preprocessor output of header files, which
may in turn result in file rebuilds.

Due to the incremental way in which these tests may be run over a
patchset this can result in inaccurate incumbent warning counts for
patches, other than the first patch in a series for which a baseline
build already occurs. This can, in turn, lead to false positives
about new warnings being introduced.

Some discussion of this problem can be found at:
- Re: [PATCH v3 08/12] testing: net-drv: add basic shaper test
  https://lore.kernel.org/netdev/20240808122042.GA3067851@kernel.org/

While possibly heavy handed, the solution in this patch does seem
address the problem. And can always be improved upon later.

Signed-off-by: Simon Horman <horms@kernel.org>