brancher: pre-generate deltas when creating a branch
Since we batch patches for testing when something fails we need
to figure out what went wrong. Instead of having to fetch the branches
locally and manually run cidiff just generate the output after creating
each branch. Link to it from the status page. Hardcoding part of the URL
isn't nice but not sure what to do about that. We could put the link
to the cidiff in the branch manifest. But it feels a little auxiliary.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>