ui: status: show branch row before results
We used to show the row with the branch name and date as the last
row, after all results. This was because the branch row was serving
as a summary line, with total number of tests.

For a while now we've had a separate, dedicated summary row.
Branch row can be first, branch row and summary line "sandwich"
the results. This seems more readable.

Since the summary row now inserts the spacing we can remove
the CSS handling for the branch row.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>