contest: vm: capture code coverage
Code coverage is a valuable info to get to know how much we can trust a test suite, and easily find out what needs to be improved.

It is quite easy to get such info with the kernel:

- The kernel needs to be compiled with GCOV_KERNEL=y, and have either
  GCOV_PROFILE_ALL=y, or GCOV_PROFILE := y set in the Makefiles. The
  recommended way is to add 'GCOV_PROFILE' in net/Makefile and
  drivers/net/Makefile.

- Before stopping the VM, the LCOV file can be captured using the 'lcov'
  tool, version >= 2.0 is recommended.

- 'genhtml' from the LCOV project can be used later to generate an HTML
  version using all the .lcov files. It could be done per LCOV file, but
  that will then only show the coverage per VM, not the global one.

This GCOV support is disabled by default. It can be enabled via
'vm.gcov=on'. I suggest to keep it off by default, and switch it to on
later when everything is in place.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>