contest: subcases: handle retry cases
In case of failure, some tests are restarted. That's great, but when
parsing the subcases (nested tests), the previous results were
overridden. Also, when displaying the results, the 'retry' field from
the main test was re-used in each subcases: quite confusing for the
reader.

Now, in case of retry, the previous 'results' list will be modified to
add a new 'retry' entry for each test that has been re-validated. If the
previous test with the same name cannot be found, that could be because
there was major issue before and some subcases have not been executed
(or the names are not fixed). In this case, a new entry with a skipped
first attempt will be added to the list. When querying the subcases
results, the new 'retry' entry will be used, or none if it is not there.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>