tree_match: handle bpf-next/net
For a few months now [1], BPF patches that are "Networking related"
should have the "bpf-next/net" prefix.

With the previous order, the PW poller was considering this prefix as
"for net", leading to these results on Patchwork [2]:

   Context              |  Check  |  Description
  ----------------------|---------|------------------------------
  netdev/tree_selection | success | Clearly marked for net
  netdev/apply          | fail    | Patch does not apply to net-0

By changing the order of the list, such patches with "[bpf-next/net]"
will be marked for "bpf-next", and they will not be applied locally like
it was the case before with "bpf-next" patches.

Link: https://lore.kernel.org/netdev/CAADnVQJgwGh+Jf=DUFuX28R2bpWVezigQYObNoKJT8UbqekOHA@mail.gmail.com/ [1]
Link: https://patchwork.kernel.org/project/netdevbpf/patch/c02fda3177b34f9e74a044833fda9761627f4d07.1730338692.git.tanggeliang@kylinos.cn/ [2]
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>