contest: fetcher: trim the result manifest at 500 entries
NIPA CI has now run for a year, we have generated ~3150 branches.
The size of the each is around 0.5MB. It's not a problem, yet,
but keep the growth in check, trip the manifest at 500 entries.

The result-fetcher does not care about disappearing entries,
all-results.json will shrink accordingly but that's only for
the best.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>