contest: fetcher: compute test stability
As the first step for HW test ingest we need to auto-judge if
given test is stable and passing or not supported by HW.

Create a DB of all test cases we have seen and compute pass/fail
statistics. We will judge a test as stable / passing if it had
15 clean runs. "undoing" the passing status will be manual
for now.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>