contest: fetcher: add autoignore based on device presence
Device driver may simply not support certain features,
so tests will be permanently failing. We want to differentiate
the tests which are broken for those not supported by a device
/ platform for ease of tracking which tests can / should be fixed.

Add a "autoignore" column (not a great name) which will
be false for SW tests but we can set to true for certain
runner + test combinations.

We expect HW results to come with a new "device" key,
use its presence to set the autoignore.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>