contest: collector: insert stability first
We want to make sure that new (HW) tests are reported as unstable
until we know they are stable. Insert stability before we break
down the data for main result insert.

It was initially late just to make sure if we crash we already
inserted the results. We have been populating the stability data
for about a week now, the code should be solid enough.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>