pw_contest: limit the number of updates on the first run
We seem to report 12 checks for the first branch that contains
a patch, because the test case count grows slowly. This happens
mostly for branches which have failures, as failure is declared
as soon as one remote fails, even if others are pending.

So limit the updates of test count for otherwise equivalent state
if some remotes are still running.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>