contest: loadavg: wait for dirty page writeback
It looks like AWS machines stall very painfully for some time
after the builds because of pending disk IO. The b/w cap was
increased on our instances but still there is a few minutes
of lag between when we decide the that the load has gone
down and when machine has recovered from the builds.

Wait for dirty memory to go down to ~100MB.

Since we have two conditions now lower the stable count to 3.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>