tests: use correct HEAD to list files
When checking the files before the patch, after a checkout to 'HEAD~',
it is required to use the previous HEAD, not the new one to look at the
same files and not others.

While at it, always use the $HEAD variable, and also fix a related
comment + added an extra one to explain the diff.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>