tests: shellcheck: keep files compliant
If a file was shellcheck compliant before or is new, it is interesting
to tell the reviewers when this status changes or if the new file is not
shellcheck compliant from the beginning.

A new dedicated file is used for each modified shell script, using a
hash of the file path to cope with files with the same base name. If the
status is different than before, warnings + info + style are added to
the errors count.

Note: it is not clear if the log message about a file no longer being
shellcheck compliant any more should be more visible or not, i.e. sent
to $DESC_FD.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>