tests: pylint/shellcheck/yamllint: shellcheck compliant
It sounds good to have the script checking for shellcheck compliance to
be shellcheck compliant :)

While at it, pylint and yamllint scripts have also be modified because
they are inspired by the shellcheck one.

Before the modifications, shellcheck was complaining about:

     53  note: Double quote to prevent globbing and word splitting. [SC2086]
      4  warning: Quote this to prevent word splitting. [SC2046]
      3  error: Argument mixes string and array. Use * or separate argument. [SC2145]
      2  warning: Use 'cd ... || exit' or 'cd ... || return' in case cd fails. [SC2164]

Not any more.

Also fix the indentation around the 'diff' at the end of each script,
and in shellcheck.sh, s/Building/Checking/ like the others.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>