tests: lint: print versions
The linters might show different results depending on the version.

Then, it looks interesting to print these versions to help people
reproducing issues on their side.

If there is an issue to print the version, exit with an error: this will
also help to identify if the linter is missing or if there is something
wrong with it.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>