tests: shellcheck: print full path in logs
It is clearer, and helpful when multiple files have the same base name.

Even more when we only have the SHA to print.

To get that, we can simply use an associative array variable.

While at it, move some code out of the sub-shells, the same way before
and after the patch, to only use these sub-shells when changing
directories.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>