contest: collector: fix input for marking as passing from srk to cur
We should only mark test as passing based on cur[rent streak]
not the [max] s[t]r[ea]k. Because if a test is incorrectly
considered as passing we want to be able to force is back
to filtered. But unless we also clear the max streak it
will come back to passing on the next collection.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>