ingest_mdir: use current branch in the kernel tree
Instead of letting user specify the tree branch use what's
checked out. This is slightly less convenient, but it will
be less error prone. It will avoid accidental changes
to the state of the tree.

Detach when we start testing to avoid overriding the branch.
Then return to the branch at the end, discarding the tested
patches.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>