contest: remove: add support for env checks
Driver tests seem to fail quite a lot due to other tests not cleaning
up after themselves. Add a hook point and a script for checking that
the env is sane. The expectation is that we'll run
contest/scripts/env_check.py in each netns before the test (as part
of setup), and then after each test. If post-test check fails VM
is destroyed and new one gets started.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>