contest: store pending results in a separate table
We used to store an entry with NULL results in the main result
table for runners which reported that they kicked off but haven't
finished testing, yet.

This was convenient for the status UI which wants to show results
for last 10 runs, and unfinished one show up as "pending".

However, mixing the NULLs in the main table makes using json
helpers in PostreSQL very annoying as we need to filter them out.
Move the pending results to their own table.

Add a option to query interface to add them in as in practice it's
only the status UI that wants to show pending runners.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>