contest: backend: use branch_date to trim results
For status UI we want to show results from last 10 branches.
This is slightly tricky as we have a row for each runner, so
there are multiple rows per branch.

We used to try to estimate how many rows we need to match 10 branches
exactly. Relying on LIMIT to separate branches doesn't work very well
because we have multiple branch streams. The rows from two branches may
be interleaved.

Use matching on branch_date instead.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>