tests: kdoc: improve comparison and output
Currently, the kdoc test fails when the number of lines of outout from
./scripts/kernel-doc increases. This is a decent heuristic but can fail in
some cases. For example, if a patch adds one warning but removes another,
the total line count may not change. This could result in false negatives
where some patches do not fail the test despite adding new warnings.

Upon detecting a difference in warning count, the test uses diff to compare
the old and new warnings. This output is not very useful to humans. The
diff comparison includes the line numbers which are impacted by code
movement, making relevant warnings difficult to spot when reviewing test
output.

An ideal kdoc test should determine the actual set of new warnings. To
achieve this, use the following algorithm:

1) Capture the kernel-doc output for the old commit
2) Capture the kernel-doc output for the new commit
3) Carefully join and sort output so that its stable w.r.t line number changes
4) Strip line numbers from the warning content for both old and new output
5) Diff the stripped output, using --*-group-format to just print line
   number ranges.
6) Use the difference line ranges to print from the unstripped output,
   generating a set of new and removed warnings.

By ensuring that the sorted input is stable even if line numbers change,
this correctly locates all of the new warnings, and additionally any
removed warnings, avoiding a false test pass.

In addition, the output is far more useful for humans as the output is 1)
just the new warnings, and doesn't contain diff syntax and 2) doesn't
contain warnings whose only change is the line number.

This is a bit more complicated, especially since it requires reading
through the input and sorting multiple times. However, the end result is
significantly more useful.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>