tests: kdoc: re-implement in python
The existing kdoc test is written in shell, based on a simple heuristic of
checking whether the number of lines of output from scripts/kernel-doc
increases.

This test can pass in some cases even when new warnings are introduced,
such as if a single patch adds some new warnings while removing others. In
addition, the provided log output is not very useful to humans. It uses a
traditional diff to compare the prior and current warnings. Because the
warnings include line number data, this results in an unreadable mess.

Implementing a proper comparison in the shell script is possible, but the
resulting logic and code is difficult to maintain. Ultimately, what we want
to do is a sort of set intersection on the warnings while ignoring line
numbers. Implementing this in python is much more practical.

As a first step, convert the kdoc test into a python module. Commands are
executed via subprocess. The improved comparison algorithm will be
implemented in a following change.

Since we will be implementing a new comparison algorithm, do not bother
porting the "diff" implementation forward. This results in a temporary
regression as we no longer compute the difference or file summary for the
user. This will be resolved with the following change implementing the
improved algorithm.

This version has a couple of minor changes worth nothing:

1) The log output is now in stdout instead of stderr, since the
   python-based tests do not support logging output to the stderr file.
2) We no longer print out "Checking the tree..." lines.
3) The description line is included within the stdout log text as well as
   the desc file.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>