ingest_mdir: fix --list-tests handling
There are 2 problems with the --list-tests command:

- The arguments --tree and the --patch/--mdir group are marked as
  required=True, forcing argparse to demand them even when they aren't
  needed (like for --list-tests).

- The main() function tries to access args.tree before checking if
  --list-tests was specified, which would cause a crash even if the
  parser requirement was removed. This problem is masked by the first
  one, so it makes sense to handle them together.

$ ./ingest_mdir.py --list-tests # should have succeeded
ingest_mdir.py: error: the following arguments are required: --tree

$ ./ingest_mdir.py --tree dummy --list-tests
ingest_mdir.py: error: one of the arguments --patch --mdir is required

$ ./ingest_mdir.py --tree dummy --patch dummy2 --list-tests # only this variant succeeds

Remove the required=True from the patch_arg group and the --tree
argument, and add manual handling for them if --list-tests wasn't
specified.

Also, move the args.tree dereference after the list_tests check and
the manual verification that the argument was provided.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>