ingest_mdir: reject invalid --test and --disable-test values
Currently, if we specify an invalid test name such as "checkpatch"
instead of "patch/checkpatch", the script falls through with no test
run, and no indication as to what was wrong. That is admittedly not very
friendly.

We'd need to call list_tests regardless of whether the --list-tests
argument was specified, and validate against its output. However, this
makes it redundant to have that function, we can just move it into main().

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>