tests: kdoc: fix parse failure logging
Commit 0a1066f7b7dd ("tests: kdoc: fix handling file removal") added extra
logging to the test intending to help debug parsing failures.

Unfortunately, the newly added "<parse fail>" log line is added to every
warning which doesn't end in ':', regardless of whether or not the
KdocWarning class handles parsing of the line.

This appears to have happened due to the confusion of the way skip and
extra work. What we really want to do is convert the text into a warning
object, then check if it cleanly parsed. We could check for the 'Unknown'
kind. However, in the interest of being precise should new kinds ever be
added, expose a new 'parsed' field of the warning object. Its set to true
if we got a clean regex parse, and false otherwise.

Additionally refactor the conditional checks and add comments for clarity.
In particular, set extra to None unconditionally first, then perform the
merging check. This should improve the readability and intended flow of
this logic.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>